feat: Found a more efficient algorithm with apoc.
