+++ config.py
@@ -118,7 +118,7 @@ link_pred_config = Config(
     p_dropout_edges=0.2,  # Currently not being used!
     p_dropout_features=0.3,
     batch_norm=True,
-    neo4j=False,
+    neo4j=True,
 )
 
 
@@ -159,6 +159,6 @@ preprocessing_config = PreprocessingConfig(
     load_text_embedding=False,
     text_embedding_colname="derived_look",
     data_size=10_000,
-    save_to_neo4j=False,
+    save_to_neo4j=True,
     data_type=DataType.pyg,
 )
